fixed editing major issue
Before I couldn't figure out to have the student's already selected major preselected in the list of the edit information page. 

INS T E A D since that was way too hard for not much extra convenience, I just have it set up so that if the student doesn't select a major in the list, it just doesn't change it. If a value is selected then it selects that major. It does NOT add the major to their existing selected, it REPLACES IT. If the student is now doing a double major, it has to select BOTH majors

I still like the idea of having it preselected but it was giving me trouble. This works and is convenient enough.